Implementada página de inicio con eventos aleatorios

+++ GR82.15.Practica1/src/es/uc3m/tiw/controller/DropOutHandler.java
@@ -1,7 +1,10 @@
 package es.uc3m.tiw.controller;
 
 import java.io.IOException;
+import java.util.List;
 
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -15,13 +18,27 @@ public class DropOutHandler implements IRequestHandler{
 			throws ServletException, IOException {
 		
 		Usr user = (Usr)request.getSession().getAttribute("loggedUser");
-		UsrDAO userDAO = new UsrDAO();
 		
-		userDAO.deleteUser(user);
+		EntityManagerFactory factory = Persistence.createEntityManagerFactory("GR82.15.Practica1");
 		
-		request.setAttribute("dropOutSuccess", true);
-		request.getSession().invalidate();
-		return "index.jsp";
+		EventManager manager = new EventManager();
+		manager.setEntityManagerFactory(factory);
+		
+		List<Event> availableEvents = manager.findAvailableEventsByCreator(user);
+		
+		if(!availableEvents.isEmpty()){
+			request.setAttribute("dropOutSuccess", false);
+			return "editProfile.jsp";
+		}
+		else {
+			UsrDAO userDAO = new UsrDAO();
+			userDAO.deleteUser(user);
+			
+			request.setAttribute("dropOutSuccess", true);
+			request.getSession().invalidate();
+			return "/index";
+		}
+		
 	}
 
 }
